Next step of the internal code changes * global: better error handling management, more robust * added Target and Client states (shared, defined in Lustre/Target.py) * added Command return codes definitions and implementations for the current commands * factorized codes in Commands * re-added fs.remove() support * re-added tune() after start() * re-propagated tuning.conf without InstallTuningConf * moved Commands.RemoteCallEventHandler to Commands.Base.RemoteCallEventHandler * unrestricted shine syntax: shine [options ...] <command> [options ...] * added (simple) modprobe before mounting targets and clients
SVN commit: r92

+++ lib/Shine/Configuration/FileSystem.py
@@ -374,10 +374,11 @@ class FileSystem(Model):
         This function aims to remove a file system configuration from
         the backend.        
         """
+        result = 0
         if self._start_backend():
             result = self.backend.unregister_fs(self)
 
-        # remove the XMF file
-        os.unlink(self.xmf_path)
+        if not result:
+            os.unlink(self.xmf_path)
 
         return result
